♻refactor: #S10P21C101-backend 평점 스코어 수정

+++ backend/src/main/java/com/ssafy/backend/domain/book/entity/BookReview.java
@@ -24,7 +24,7 @@ public class BookReview {
     @ManyToOne
     private Book book;
 
-    private int score;
+    private double score;
 
     private String content;
